Sync working tree across impulse, controller, bilateral, utils
Broad checkpoint of in-progress work (figures/data gitignored):
- impulse-analysis: contra_residual.m (residual-only pipeline), plus
  dose_response/run_all/spatial_spread/tf_fit updates
- controller-analysis: trial_state_mse, spiral_analysis, widebrain_rrr,
  widebrain_hemi_kernel, pixelviewer, run_all; refactors to
  cl_mse_factors/load_sessions/motion_analysis/step_response/tf_fit;
  retire prestim_variance.m + spectral_mse_sort.m
- bilateral-analysis/ and explore/ scaffolding added
- utils: redoSVD.m + helper updates
- docs: MEETINGS.md, sfn_abstract drafts, data_overview.m

+++ utils/initialize_data_nosvd.m
@@ -43,7 +43,7 @@ function d = initialize_data_nosvd(mn,en,td)
     % px=[];
     % py=[];
     
-    frame = double([d.params.pixel(1),px;d.params.pixel(2),py]);
+    frame = [double(d.params.pixel(1)),px;double(d.params.pixel(2)),py];
     d.params.pixels = frame';
